refactor: Transaction 처리하는 로직 분리
