+++ app/src/main/java/com/techcourse/service/UserService.java
@@ -1,6 +1,7 @@
 package com.techcourse.service;
 
 import com.interface21.dao.DataAccessException;
+import com.interface21.jdbc.core.TransactionManager;
 import com.techcourse.config.DataSourceConfig;
 import com.techcourse.dao.UserDao;
 import com.techcourse.dao.UserHistoryDao;
@@ -20,11 +21,13 @@ public class UserService {
     private final UserDao userDao;
     private final UserHistoryDao userHistoryDao;
     private final DataSource dataSource;
+    private final TransactionManager transactionManager;
 
     public UserService(final UserDao userDao, final UserHistoryDao userHistoryDao) {
         this.userDao = userDao;
         this.userHistoryDao = userHistoryDao;
         this.dataSource = DataSourceConfig.getInstance();
+        this.transactionManager = new TransactionManager();
     }
 
     public User findById(final long id) {
@@ -45,30 +48,11 @@ public class UserService {
     }
 
     private void updatePassword(final Connection connection, final long id, final String newPassword, final String createBy) {
-        try {
-            connection.setAutoCommit(false);
-
+        transactionManager.start(connection, () -> {
             final var user = findById(id);
             user.changePassword(newPassword);
             userDao.update(connection, user);
             userHistoryDao.log(connection, new UserHistory(user, createBy));
-
-            connection.commit();
-        } catch (SQLException e) {
-            log.error(e.getMessage(), e);
-            rollback(connection);
-            throw new DataAccessException(e);
-        }
-    }
-
-    private void rollback(final Connection connection) {
-        if (connection != null) {
-            try {
-                connection.rollback();
-            } catch (SQLException e) {
-                log.error(e.getMessage(), e);
-                throw new DataAccessException(e);
-            }
-        }
+        });
     }
 }
